requirements.txt
ttkbootstrap>=1.10.1
yt-dlp>=2023.0.0
Pillow>=10.0.0
plyer>=2.1.0
duckduckgo-search>=6.1.0
static-ffmpeg>=3.0.0
subliminal>=2.1.0
